# services/gateway/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EcommerceGo API Gateway",
    "description": "Unified entry point for all EcommerceGo microservices. All requests require a JWT Bearer token in the Authorization header unless noted otherwise. Prices are always in cents (int64).",
    "version": "1.0.0",
    "contact": {
      "name": "EcommerceGo",
      "url": "https://github.com/utafrali/EcommerceGo"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development (API Gateway)"
    }
  ],
  "security": [
    { "BearerAuth": [] }
  ],
...
  "paths": {
...
    "/api/v1/users/{id}/profile": {
      "get": {
        "tags": ["users"],
        "summary": "Get user profile",
        "operationId": "getProfile",
        "parameters": [ { "$ref": "#/components/parameters/IDParam" } ],
        "responses": {
          "200": { "description": "User profile" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "tags": ["users"],
        "summary": "Update user profile",
        "operationId": "updateProfile",
        "parameters": [ { "$ref": "#/components/parameters/IDParam" } ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateProfileRequest" } } } },
        "responses": {
          "200": { "description": "Profile updated" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/users/{id}/addresses": {
      "get": {
        "tags": ["users"],
        "summary": "List user addresses",
        "operationId": "listAddresses",
        "parameters": [ { "$ref": "#/components/parameters/IDParam" } ],
        "responses": {
          "200": { "description": "Address list" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["users"],
        "summary": "Create address",
        "operationId": "createAddress",
        "parameters": [ { "$ref": "#/components/parameters/IDParam" } ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AddressRequest" } } } },
        "responses": {
          "201": { "description": "Address created" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/v1/users/{id}/addresses/{addressId}": {
      "put": {
        "tags": ["users"],
        "summary": "Update address",
        "operationId": "updateAddress",
        "parameters": [
          { "$ref": "#/components/parameters/IDParam" },
          { "name": "addressId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AddressRequest" } } } },
        "responses": {
          "200": { "description": "Address updated" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["users"],
        "summary": "Delete address",
        "operationId": "deleteAddress",
        "parameters": [
          { "$ref": "#/components/parameters/IDParam" },
          { "name": "addressId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Address deleted" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /api/v1/auth/login or /api/v1/auth/register"
      }
    },
    "parameters": {
      "IDParam": {
        "name": "id",
        "in": "path",
        "description": "Resource UUID",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request body or parameters",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
      "Unauthorized": {
        "description": "Missing or invalid JWT token",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
...
    },
    "schemas": {
...
      "ErrorDetail": {
        "type": "object",
        "properties": {
          "code": { "type": "string", "example": "UNAUTHORIZED" },
          "message": { "type": "string", "example": "missing or invalid token" },
          "request_id": { "type": "string", "example": "req-abc123" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "$ref": "#/components/schemas/ErrorDetail" }
        }
      },
...
      "UpdateProfileRequest": {
        "type": "object",
        "properties": {
          "first_name": { "type": "string" },
          "last_name": { "type": "string" },
          "phone": { "type": "string" }
        }
      },
      "AddressRequest": {
        "type": "object",
        "required": ["line1", "city", "country"],
        "properties": {
          "line1": { "type": "string" },
          "line2": { "type": "string" },
          "city": { "type": "string" },
          "state": { "type": "string" },
          "postal_code": { "type": "string" },
          "country": { "type": "string", "example": "TR" },
          "is_default": { "type": "boolean" }
        }
      },
...
    }
  }
}